requirements.txt
Django==2.2
django-environ==0.9.0
djangorestframework==3.9.2
pytz==2022.1
sqlparse==0.4.2
